pyproject.toml
[project]
name = "trim2sort"
version = "0.1.0"
description = "一個用於 DNA 序列分析的圖形化使用者介面（GUI）應用程式，支援 Sanger 定序與次世代定序（NGS）資料處理與分析"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.86",
    "customtkinter>=5.2.2",
    "natsort>=8.4.0",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "jinja2>=3.1.6",
    "rpy2>=3.6.0",
]

[tool.ruff]
# 目標 Python 版本
target-version = "py312"
# 行長度限制
line-length = 100
# 排除的檔案和目錄
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# 啟用的規則集
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "RUF", # Ruff-specific rules
]
# 忽略的規則
ignore = [
    "E501",  # 行長度（由 formatter 處理）
    "B008",  # 函數參數中的函數調用（某些情況下是必要的）
    "ARG001", # 未使用的函數參數（在 GUI 中很常見）
    "ARG002", # 未使用的類方法參數
    "F821",  # 未定義名稱（動態通過 globals() 創建的變數）
    "PTH",   # pathlib 建議（保持使用 os.path 以保持兼容性）
]

[tool.ruff.lint.isort]
# 使用單行導入
force-single-line = false
# 按字母順序排序
force-sort-within-sections = true
# 已知的第一方導入
known-first-party = []

[tool.ruff.format]
# 使用雙引號
quote-style = "double"
# 縮排樣式
indent-style = "space"
# 縮排寬度
skip-magic-trailing-comma = false
# 行尾樣式
line-ending = "auto"

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]
